Create Hypot pyui
This is a .pyui file in pythonista3. I converted it to a json file so it can be readable. I want the input from ‘length’ (A) and input from ‘width’ (B) to be calculated for ‘hypot’ (C) and be presented in textview ‘Hypotenuse’ after hitting the calculate button. This is not working... I’m sure the solution is simple, but I am very green (so try to avoid judgment, we all start somewhere)